Added save/restore feature
The values of the weights between the input and the hidden layer as well as the ones between the hidden and the outer layer of the trained network are saved into a .csv file. This way the neural network can be restored to an earlier state without needing to train it again. Very useful for testing different datasets with a given network.